Makes axis text labels flush on the ends

A convenience function intended for basic, fixed-scale plots only (i.e. does not handle free scales in facets).

You need to pass in a ggplot2 object to this function. It can't be +'d in a chain of geoms, coords, scales, themes, etc. It also builds the plot (but does not display it) so if the plot takes a while (i.e. has lots of data or transforms) this will also take a while.

flush_ticks(gg, flush = "XY", plot = TRUE, cat = TRUE)

Arguments

gg

ggplot2 plot object

flush

either "X" or "Y" or "XY" to flush individual or both axes. Default: both.

plot

if FALSE then the ggplot object will be returned invisibly

cat

if TRUE then display theme() statements and copy them to the clipboard

Value

ggplot2 object with theme() elements added

Note

Intended for basic, fixed-scale plots only (i.e. does not handle free scales in facets).