Import Goldman Sans font for use in charts

Goldman Sans is a trademark of Goldman Sachs and distributed under the Goldman Sachs Restricted Font License.

import_goldman_sans()

Details

There is an option hrbrthemes.loadfonts which -- if set to TRUE -- will call extrafont::loadfonts() to register non-core fonts with R PDF & PostScript devices. If you are running under Windows, the package calls the same function to register non-core fonts with the Windows graphics device.

Note

This will take care of ensuring PDF/PostScript usage. The location of the font directory is displayed after the base import is complete. It is highly recommended that you install them on your system the same way you would any other font you wish to use in other programs.