X & Y scales with opinionated pre-sets for percent & comma label formats

The _comma ones set comma format for axis text and expand=c(0,0) (you need to set limits).

scale_x_percent(name = waiver(), breaks = waiver(),
  minor_breaks = waiver(), labels = scales::percent, limits = NULL,
  expand = c(0.01, 0), oob = censor, na.value = NA_real_,
  trans = "identity", position = "bottom", sec.axis = waiver())

scale_y_percent(name = waiver(), breaks = waiver(),
  minor_breaks = waiver(), labels = scales::percent, limits = NULL,
  expand = c(0.01, 0), oob = censor, na.value = NA_real_,
  trans = "identity", position = "left", sec.axis = waiver())

scale_x_comma(name = waiver(), breaks = waiver(), minor_breaks = waiver(),
  labels = scales::comma, limits = NULL, expand = c(0.01, 0),
  oob = censor, na.value = NA_real_, trans = "identity",
  position = "bottom", sec.axis = waiver())

scale_y_comma(name = waiver(), breaks = waiver(), minor_breaks = waiver(),
  labels = scales::comma, limits = NULL, expand = c(0.01, 0),
  oob = censor, na.value = NA_real_, trans = "identity",
  position = "left", sec.axis = waiver())

Arguments

name
The name of the scale. Used as axis or legend title. If NULL, the default, the name of the scale is taken from the first mapping used for that aesthetic.

breaks
One of:

NULL for no breaks

waiver() for the default breaks computed by the transformation object

A numeric vector of positions

A function that takes the limits as input and returns breaks as output

minor_breaks
One of:

NULL for no minor breaks

waiver() for the default breaks (one minor break between each major break)

A numeric vector of positions

A function that given the limits returns a vector of minor breaks.

labels
One of:

NULL for no labels

waiver() for the default labels computed by the transformation object

A character vector giving labels (must be same length as breaks)

A function that takes the breaks as input and returns labels as output

limits
A numeric vector of length two providing limits of the scale. Use NA to refer to the existing minimum or maximum.

expand
A numeric vector of length two giving multiplicative and additive expansion constants. These constants ensure that the data is placed some distance away from the axes. The defaults are c(0.05, 0) for continuous variables, and c(0, 0.6) for discrete variables.

oob
Function that handles limits outside of the scale limits (out of bounds). The default replaces out of bounds values with NA.

na.value
Missing values will be replaced with this value.

trans
Either the name of a transformation object, or the object itself. Built-in transformations include "asn", "atanh", "boxcox", "exp", "identity", "log", "log10", "log1p", "log2", "logit", "probability", "probit", "reciprocal", "reverse" and "sqrt".

A transformation object bundles together a transform, it's inverse, and methods for generating breaks and labels. Transformation objects are defined in the scales package, and are called name_trans, e.g. boxcox_trans. You can create your own transformation with trans_new.

position
The position of the axis. "left" or "right" for vertical scales, "top" or "bottom" for horizontal scales

sec.axis
specifify a secondary axis

Details

The _percent ones set precent format for axis text and expand=c(0,0) (you need to set limits).